Fix formatting of storage files
